feat(Aliases): <type>d should refer to 'double'

+++ src/math/rect.rs
@@ -7,7 +7,7 @@ use crate::math::{Scalar, Unit, Vector, Vector2};
 pub type Rect2u<U> = Rect2<u32, U>;
 pub type Rect2i<U> = Rect2<i32, U>;
 pub type Rect2f<U> = Rect2<f32, U>;
-pub type Rect2d<U> = Rect2<i64, U>;
+pub type Rect2d<U> = Rect2<f64, U>;
 
 #[derive(Debug, Default, Clone, Copy)]
 pub struct Rect2<S, U = ()>
